Refactor ingest path: fix unit tests due to change of absolute paths.
