+++ pkg/pgmodel/ingester/ingestor_test.go
@@ -1,96 +1,11 @@
-// This file and its contents are licensed under the Apache License 2.0.
-// Please see the included NOTICE for copyright information and
-// LICENSE for a copy of the license.
-package utils
+package ingester
 
 import (
 	"fmt"
-	"github.com/timescale/promscale/pkg/pgmodel/cache"
-	"github.com/timescale/promscale/pkg/pgmodel/ingester"
-	"testing"
-
+	"github.com/timescale/promscale/pkg/pgmodel/utils"
 	"github.com/timescale/promscale/pkg/prompb"
+	"testing"
 )
-
-type mockCache struct {
-	seriesCache  map[string]SeriesID
-	getSeriesErr error
-	setSeriesErr error
-}
-
-var _ SeriesCache = (*mockCache)(nil)
-
-func (m *mockCache) GetSeries(lset Labels) (SeriesID, error) {
-	if m.getSeriesErr != nil {
-		return 0, m.getSeriesErr
-	}
-
-	val, ok := m.seriesCache[lset.String()]
-	if !ok {
-		return 0, cache.ErrEntryNotFound
-	}
-
-	return val, nil
-}
-
-func (m *mockCache) SetSeries(lset Labels, id SeriesID) error {
-	m.seriesCache[lset.String()] = id
-	return m.setSeriesErr
-}
-
-func (m *mockCache) NumElements() int {
-	return len(m.seriesCache)
-}
-
-func (m *mockCache) Capacity() int {
-	return len(m.seriesCache)
-}
-
-type mockInserter struct {
-	insertedSeries  map[string]SeriesID
-	insertedData    []map[string][]samplesInfo
-	insertSeriesErr error
-	insertDataErr   error
-}
-
-func (m *mockInserter) Close() {
-
-}
-
-func (m *mockInserter) InsertNewData(rows map[string][]samplesInfo) (uint64, error) {
-	return m.InsertData(rows)
-}
-
-func (m *mockInserter) CompleteMetricCreation() error {
-	return nil
-}
-
-func (m *mockInserter) InsertData(rows map[string][]samplesInfo) (uint64, error) {
-	for _, v := range rows {
-		for i, si := range v {
-			id, ok := m.insertedSeries[si.labels.String()]
-			if !ok {
-				id = SeriesID(len(m.insertedSeries))
-				m.insertedSeries[si.labels.String()] = id
-			}
-			v[i].seriesID = id
-		}
-	}
-	if m.insertSeriesErr != nil {
-		return 0, m.insertSeriesErr
-	}
-	m.insertedData = append(m.insertedData, rows)
-	ret := 0
-	for _, data := range rows {
-		for _, si := range data {
-			ret += len(si.samples)
-		}
-	}
-	if m.insertDataErr != nil {
-		ret = 0
-	}
-	return uint64(ret), m.insertDataErr
-}
 
 func TestDBIngestorIngest(t *testing.T) {
 	testCases := []struct {
@@ -112,7 +27,7 @@ func TestDBIngestorIngest(t *testing.T) {
 			metrics: []prompb.TimeSeries{
 				{
 					Labels: []prompb.Label{
-						{Name: MetricNameLabelName, Value: "test"},
+						{Name: utils.MetricNameLabelName, Value: "test"},
 					},
 					Samples: []prompb.Sample{
 						{Timestamp: 1, Value: 0.1},
@@ -127,7 +42,7 @@ func TestDBIngestorIngest(t *testing.T) {
 			metrics: []prompb.TimeSeries{
 				{
 					Labels: []prompb.Label{
-						{Name: MetricNameLabelName, Value: "test"},
+						{Name: utils.MetricNameLabelName, Value: "test"},
 						{Name: "test", Value: "test"},
 					},
 				},
@@ -138,7 +53,7 @@ func TestDBIngestorIngest(t *testing.T) {
 			metrics: []prompb.TimeSeries{
 				{
 					Labels: []prompb.Label{
-						{Name: MetricNameLabelName, Value: "test"},
+						{Name: utils.MetricNameLabelName, Value: "test"},
 						{Name: "foo", Value: "bar"},
 					},
 					Samples: []prompb.Sample{
@@ -147,7 +62,7 @@ func TestDBIngestorIngest(t *testing.T) {
 				},
 				{
 					Labels: []prompb.Label{
-						{Name: MetricNameLabelName, Value: "test"},
+						{Name: utils.MetricNameLabelName, Value: "test"},
 						{Name: "test", Value: "test"},
 					},
 					Samples: []prompb.Sample{
@@ -163,7 +78,7 @@ func TestDBIngestorIngest(t *testing.T) {
 			metrics: []prompb.TimeSeries{
 				{
 					Labels: []prompb.Label{
-						{Name: MetricNameLabelName, Value: "test"},
+						{Name: utils.MetricNameLabelName, Value: "test"},
 						{Name: "test", Value: "test"},
 					},
 					Samples: []prompb.Sample{
@@ -180,7 +95,7 @@ func TestDBIngestorIngest(t *testing.T) {
 			metrics: []prompb.TimeSeries{
 				{
 					Labels: []prompb.Label{
-						{Name: MetricNameLabelName, Value: "test"},
+						{Name: utils.MetricNameLabelName, Value: "test"},
 						{Name: "test", Value: "test"},
 					},
 					Samples: []prompb.Sample{
@@ -189,7 +104,7 @@ func TestDBIngestorIngest(t *testing.T) {
 				},
 				{
 					Labels: []prompb.Label{
-						{Name: MetricNameLabelName, Value: "test"},
+						{Name: utils.MetricNameLabelName, Value: "test"},
 						{Name: "test", Value: "test"},
 					},
 					Samples: []prompb.Sample{
@@ -205,7 +120,7 @@ func TestDBIngestorIngest(t *testing.T) {
 			metrics: []prompb.TimeSeries{
 				{
 					Labels: []prompb.Label{
-						{Name: MetricNameLabelName, Value: "test"},
+						{Name: utils.MetricNameLabelName, Value: "test"},
 						{Name: "test", Value: "test"},
 					},
 					Samples: []prompb.Sample{
@@ -222,7 +137,7 @@ func TestDBIngestorIngest(t *testing.T) {
 			metrics: []prompb.TimeSeries{
 				{
 					Labels: []prompb.Label{
-						{Name: MetricNameLabelName, Value: "test"},
+						{Name: utils.MetricNameLabelName, Value: "test"},
 						{Name: "test", Value: "test"},
 					},
 					Samples: []prompb.Sample{
@@ -250,17 +165,17 @@ func TestDBIngestorIngest(t *testing.T) {
 
 	for _, c := range testCases {
 		t.Run(c.name, func(t *testing.T) {
-			inserter := mockInserter{
-				insertSeriesErr: c.insertSeriesErr,
-				insertDataErr:   c.insertDataErr,
-				insertedSeries:  make(map[string]SeriesID),
+			inserter := utils.MockInserter{
+				InsertSeriesErr: c.insertSeriesErr,
+				InsertDataErr:   c.insertDataErr,
+				InsertedSeries:  make(map[string]utils.SeriesID),
 			}
 
-			i := ingester.DBIngestor{
+			i := DBIngestor{
 				db: &inserter,
 			}
 
-			count, err := i.Ingest(c.metrics, ingester.NewWriteRequest())
+			count, err := i.Ingest(c.metrics, NewWriteRequest())
 
 			if err != nil {
 				if c.insertSeriesErr != nil && err != c.insertSeriesErr {
@@ -275,10 +190,10 @@ func TestDBIngestorIngest(t *testing.T) {
 				if c.setSeriesErr != nil && err != c.setSeriesErr {
 					t.Errorf("wrong error returned: got\n%s\nwant\n%s\n", err, c.setSeriesErr)
 				}
-				if err == ErrNoMetricName {
+				if err == utils.ErrNoMetricName {
 					for _, ts := range c.metrics {
 						for _, label := range ts.Labels {
-							if label.Name == MetricNameLabelName {
+							if label.Name == utils.MetricNameLabelName {
 								t.Errorf("returning missing metric name when one was found for metric name: %s\n", label.Name)
 							}
 						}
@@ -290,11 +205,11 @@ func TestDBIngestorIngest(t *testing.T) {
 				t.Errorf("invalid number of metrics inserted: got %d, want %d\n", count, c.count)
 			}
 
-			if c.countSeries != len(inserter.insertedSeries) {
+			if c.countSeries != len(inserter.InsertedSeries) {
 				t.Errorf("invalid number of series inserted, all series that are not cached must be sent for insertion: got %d, want %d\n%+v\n",
-					len(inserter.insertedSeries),
+					len(inserter.InsertedSeries),
 					c.countSeries,
-					inserter.insertedSeries,
+					inserter.InsertedSeries,
 				)
 			}
 		})
