+++ pkg/pgmodel/sql_test.go
@@ -25,373 +25,6 @@ import (
 	"github.com/timescale/promscale/pkg/prompb"
 )
 
-type sqlRecorder struct {
-	queries   []sqlQuery
-	nextQuery int
-	lock      sync.Mutex
-	t         *testing.T
-}
-
-type sqlQuery struct {
-	sql     string
-	args    []interface{}
-	results rowResults
-	err     error
-}
-
-// rowResults represents a collection of a multi-column row result
-type rowResults [][]interface{}
-
-func newSqlRecorder(queries []sqlQuery, t *testing.T) *sqlRecorder {
-	return &sqlRecorder{queries: queries, t: t}
-}
-
-func (r *sqlRecorder) Close() {
-}
-
-func (r *sqlRecorder) Exec(ctx context.Context, sql string, arguments ...interface{}) (pgconn.CommandTag, error) {
-	r.lock.Lock()
-	defer r.lock.Unlock()
-
-	results, err := r.checkQuery(sql, arguments...)
-
-	if len(results) == 0 {
-		return nil, err
-	}
-	if len(results) != 1 {
-		r.t.Errorf("mock exec: too many return rows %v\n in Exec\n %v\n args %v",
-			results, sql, arguments)
-		return nil, err
-	}
-	if len(results[0]) != 1 {
-		r.t.Errorf("mock exec: too many return values %v\n in Exec\n %v\n args %v",
-			results, sql, arguments)
-		return nil, err
-	}
-
-	return results[0][0].(pgconn.CommandTag), err
-}
-
-func (r *sqlRecorder) Query(ctx context.Context, sql string, args ...interface{}) (pgx.Rows, error) {
-	r.lock.Lock()
-	defer r.lock.Unlock()
-	rows, err := r.checkQuery(sql, args...)
-	return &mockRows{results: rows}, err
-}
-
-func (r *sqlRecorder) QueryRow(ctx context.Context, sql string, args ...interface{}) pgx.Row {
-	r.lock.Lock()
-	defer r.lock.Unlock()
-	rows, err := r.checkQuery(sql, args...)
-	return &mockRows{results: rows, err: err}
-}
-
-func (r *sqlRecorder) CopyFrom(ctx context.Context, tableName pgx.Identifier, columnNames []string, rowSrc pgx.CopyFromSource) (int64, error) {
-	panic("should never be called")
-}
-
-func (r *sqlRecorder) CopyFromRows(rows [][]interface{}) pgx.CopyFromSource {
-	panic("should never be called")
-}
-
-func (r *sqlRecorder) NewBatch() pgxconn.PgxBatch {
-	return &mockBatch{}
-}
-
-func (r *sqlRecorder) SendBatch(ctx context.Context, b pgxconn.PgxBatch) (pgx.BatchResults, error) {
-	r.lock.Lock()
-	defer r.lock.Unlock()
-	batch := b.(*mockBatch)
-
-	start := r.nextQuery
-	for _, q := range batch.items {
-		_, _ = r.checkQuery(q.query, q.arguments...)
-	}
-	// TODO switch to q.query[] subslice
-	return &mockBatchResult{queries: r.queries[start:r.nextQuery]}, nil
-}
-
-func (r *sqlRecorder) checkQuery(sql string, args ...interface{}) (rowResults, error) {
-	idx := r.nextQuery
-	if idx >= len(r.queries) {
-		r.t.Errorf("@ %d extra query: %s", idx, sql)
-		return nil, fmt.Errorf("extra query")
-	}
-	row := r.queries[idx]
-	r.nextQuery += 1
-	if sql != row.sql {
-		r.t.Errorf("@ %d unexpected query:\ngot:\n\t%s\nexpected:\n\t%s", idx, sql, row.sql)
-	}
-	if !reflect.DeepEqual(args, row.args) {
-		r.t.Errorf("@ %d unexpected query args for\n\t%s\ngot:\n\t%#v\nexpected:\n\t%#v", idx, sql, args, row.args)
-	}
-	return row.results, row.err
-}
-
-type batchItem struct {
-	query     string
-	arguments []interface{}
-}
-
-// Batch queries are a way of bundling multiple queries together to avoid
-// unnecessary network round trips.
-type mockBatch struct {
-	items []batchItem
-}
-
-func (b *mockBatch) Queue(query string, arguments ...interface{}) {
-	b.items = append(b.items, batchItem{
-		query:     query,
-		arguments: arguments,
-	})
-}
-
-type mockBatchResult struct {
-	idx     int
-	queries []sqlQuery
-	t       *testing.T
-}
-
-// Exec reads the results from the next query in the batch as if the query has been sent with Conn.Exec.
-func (m *mockBatchResult) Exec() (pgconn.CommandTag, error) {
-	defer func() { m.idx++ }()
-
-	q := m.queries[m.idx]
-
-	if len(q.results) == 0 {
-		return nil, q.err
-	}
-	if len(q.results) != 1 {
-		m.t.Errorf("mock exec: too many return rows %v\n in batch Exec\n %+v", q.results, q)
-		return nil, q.err
-	}
-	if len(q.results[0]) != 1 {
-		m.t.Errorf("mock exec: too many return values %v\n in batch Exec\n %+v", q.results, q)
-		return nil, q.err
-	}
-
-	return q.results[0][0].(pgconn.CommandTag), q.err
-}
-
-// Query reads the results from the next query in the batch as if the query has been sent with Conn.Query.
-func (m *mockBatchResult) Query() (pgx.Rows, error) {
-	defer func() { m.idx++ }()
-
-	q := m.queries[m.idx]
-	return &mockRows{results: q.results, noNext: false}, q.err
-}
-
-// Close closes the batch operation. This must be called before the underlying connection can be used again. Any error
-// that occurred during a batch operation may have made it impossible to resyncronize the connection with the server.
-// In this case the underlying connection will have been closed.
-func (m *mockBatchResult) Close() error {
-	return nil
-}
-
-// QueryRow reads the results from the next query in the batch as if the query has been sent with Conn.QueryRow.
-func (m *mockBatchResult) QueryRow() pgx.Row {
-	defer func() { m.idx++ }()
-	q := m.queries[m.idx]
-	return &mockRows{results: q.results, err: q.err, noNext: false}
-}
-
-type mockRows struct {
-	idx     int
-	noNext  bool
-	results rowResults
-	err     error
-}
-
-// Close closes the rows, making the connection ready for use again. It is safe
-// to call Close after rows is already closed.
-func (m *mockRows) Close() {
-}
-
-// Err returns any error that occurred while reading.
-func (m *mockRows) Err() error {
-	return m.err
-}
-
-// CommandTag returns the command tag from this query. It is only available after Rows is closed.
-func (m *mockRows) CommandTag() pgconn.CommandTag {
-	panic("not implemented")
-}
-
-func (m *mockRows) FieldDescriptions() []pgproto3.FieldDescription {
-	panic("not implemented")
-}
-
-// Next prepares the next row for reading. It returns true if there is another
-// row and false if no more rows are available. It automatically closes rows
-// when all rows are read.
-func (m *mockRows) Next() bool {
-	return !m.noNext && m.idx < len(m.results)
-}
-
-// Scan reads the values from the current row into dest values positionally.
-// dest can include pointers to core types, values implementing the Scanner
-// interface, []byte, and nil. []byte will skip the decoding process and directly
-// copy the raw bytes received from PostgreSQL. nil will skip the value entirely.
-func (m *mockRows) Scan(dest ...interface{}) error {
-	defer func() { m.idx++ }()
-
-	if m.err != nil {
-		return m.err
-	}
-
-	if m.idx >= len(m.results) {
-		return fmt.Errorf("mock scanning error, no more results in batch: got %d wanted %d", m.idx, len(m.results))
-	}
-
-	if len(dest) > len(m.results[m.idx]) {
-		return fmt.Errorf("mock scanning error, missing results for scanning: got %d %#v\nwanted %d",
-			len(m.results[m.idx]),
-			m.results[m.idx],
-			len(dest),
-		)
-	}
-
-	for i := range dest {
-		switch s := m.results[m.idx][i].(type) {
-		case []time.Time:
-			if d, ok := dest[i].(*[]time.Time); ok {
-				*d = s
-			} else if d, ok := dest[i].(*pgtype.TimestamptzArray); ok {
-				*d = pgtype.TimestamptzArray{
-					Elements: make([]pgtype.Timestamptz, len(s)),
-				}
-				for i := range s {
-					d.Elements[i] = pgtype.Timestamptz{
-						Time: s[i],
-					}
-				}
-			}
-		case []float64:
-			if d, ok := dest[i].(*[]float64); ok {
-				*d = s
-			} else if d, ok := dest[i].(*pgtype.Float8Array); ok {
-				*d = pgtype.Float8Array{
-					Elements: make([]pgtype.Float8, len(s)),
-				}
-				for i := range s {
-					d.Elements[i] = pgtype.Float8{
-						Float: s[i],
-					}
-				}
-			}
-		case []int64:
-			if d, ok := dest[i].(*[]int64); ok {
-				*d = s
-				continue
-			}
-			if d, ok := dest[i].(*[]utils.SeriesID); ok {
-				for _, id := range s {
-					*d = append(*d, utils.SeriesID(id))
-				}
-				continue
-			}
-			return fmt.Errorf("wrong value type []int64")
-		case []string:
-			if d, ok := dest[i].(*[]string); ok {
-				*d = s
-			}
-		case time.Time:
-			if d, ok := dest[i].(*time.Time); ok {
-				*d = s
-			}
-		case float64:
-			if _, ok := dest[i].(float64); !ok {
-				return fmt.Errorf("wrong value type float64")
-			}
-			dv := reflect.ValueOf(dest[i])
-			dvp := reflect.Indirect(dv)
-			dvp.SetFloat(float64(m.results[m.idx][i].(float64)))
-		case int:
-			if _, ok := dest[i].(int); !ok {
-				return fmt.Errorf("wrong value type int")
-			}
-			dv := reflect.ValueOf(dest[i])
-			dvp := reflect.Indirect(dv)
-			dvp.SetInt(int64(m.results[m.idx][i].(int32)))
-		case int32:
-			if _, ok := dest[i].(int32); !ok {
-				return fmt.Errorf("wrong value type int32")
-			}
-			dv := reflect.ValueOf(dest[i])
-			dvp := reflect.Indirect(dv)
-			dvp.SetInt(int64(m.results[m.idx][i].(int32)))
-		case uint64:
-			if _, ok := dest[i].(uint64); !ok {
-				return fmt.Errorf("wrong value type uint64")
-			}
-			dv := reflect.ValueOf(dest[i])
-			dvp := reflect.Indirect(dv)
-			dvp.SetUint(m.results[m.idx][i].(uint64))
-		case int64:
-			_, ok1 := dest[i].(*int64)
-			_, ok2 := dest[i].(*utils.SeriesID)
-			if !ok1 && !ok2 {
-				return fmt.Errorf("wrong value type int64")
-			}
-			dv := reflect.ValueOf(dest[i])
-			dvp := reflect.Indirect(dv)
-			dvp.SetInt(m.results[m.idx][i].(int64))
-		case string:
-			if _, ok := dest[i].(*string); !ok {
-				return fmt.Errorf("wrong value type string")
-			}
-			dv := reflect.ValueOf(dest[i])
-			dvp := reflect.Indirect(dv)
-			dvp.SetString(m.results[m.idx][i].(string))
-		}
-	}
-
-	return nil
-}
-
-// Values returns the decoded row values.
-func (m *mockRows) Values() ([]interface{}, error) {
-	panic("not implemented")
-}
-
-// RawValues returns the unparsed bytes of the row values. The returned [][]byte is only valid until the next Next
-// call or the Rows is closed. However, the underlying byte data is safe to retain a reference to and mutate.
-func (m *mockRows) RawValues() [][]byte {
-	panic("not implemented")
-}
-
-type mockMetricCache struct {
-	metricCache  map[string]string
-	getMetricErr error
-	setMetricErr error
-}
-
-func (m *mockMetricCache) Len() int {
-	return len(m.metricCache)
-}
-
-func (m *mockMetricCache) Cap() int {
-	return len(m.metricCache)
-}
-
-func (m *mockMetricCache) Get(metric string) (string, error) {
-	if m.getMetricErr != nil {
-		return "", m.getMetricErr
-	}
-
-	val, ok := m.metricCache[metric]
-	if !ok {
-		return "", cache.ErrEntryNotFound
-	}
-
-	return val, nil
-}
-
-func (m *mockMetricCache) Set(metric string, tableName string) error {
-	m.metricCache[metric] = tableName
-	return m.setMetricErr
-}
-
 // test code start
 
 // sql_ingest tests
