+++ pkg/pgmodel/utils/testutils.go
@@ -1,0 +1,468 @@
+package utils
+
+import (
+	"context"
+	"fmt"
+	"github.com/jackc/pgconn"
+	"github.com/jackc/pgproto3/v2"
+	"github.com/jackc/pgtype"
+	"github.com/jackc/pgx/v4"
+	"github.com/timescale/promscale/pkg/pgmodel/cache"
+	"github.com/timescale/promscale/pkg/pgxconn"
+	"reflect"
+	"sync"
+	"testing"
+	"time"
+)
+
+var (
+	// ErrEntryNotFound is returned when entry is not found.
+	ErrEntryNotFound = fmt.Errorf("entry not found")
+)
+
+type SqlRecorder struct {
+	queries   []SqlQuery
+	nextQuery int
+	lock      sync.Mutex
+	t         *testing.T
+}
+
+type SqlQuery struct {
+	sql     string
+	args    []interface{}
+	results RowResults
+	err     error
+}
+
+// RowResults represents a collection of a multi-column row result
+type RowResults [][]interface{}
+
+func NewSqlRecorder(queries []SqlQuery, t *testing.T) *SqlRecorder {
+	return &SqlRecorder{queries: queries, t: t}
+}
+
+func (r *SqlRecorder) Close() {
+}
+
+func (r *SqlRecorder) Exec(ctx context.Context, sql string, arguments ...interface{}) (pgconn.CommandTag, error) {
+	r.lock.Lock()
+	defer r.lock.Unlock()
+
+	results, err := r.checkQuery(sql, arguments...)
+
+	if len(results) == 0 {
+		return nil, err
+	}
+	if len(results) != 1 {
+		r.t.Errorf("mock exec: too many return rows %v\n in Exec\n %v\n args %v",
+			results, sql, arguments)
+		return nil, err
+	}
+	if len(results[0]) != 1 {
+		r.t.Errorf("mock exec: too many return values %v\n in Exec\n %v\n args %v",
+			results, sql, arguments)
+		return nil, err
+	}
+
+	return results[0][0].(pgconn.CommandTag), err
+}
+
+func (r *SqlRecorder) Query(ctx context.Context, sql string, args ...interface{}) (pgx.Rows, error) {
+	r.lock.Lock()
+	defer r.lock.Unlock()
+	rows, err := r.checkQuery(sql, args...)
+	return &MockRows{results: rows}, err
+}
+
+func (r *SqlRecorder) QueryRow(ctx context.Context, sql string, args ...interface{}) pgx.Row {
+	r.lock.Lock()
+	defer r.lock.Unlock()
+	rows, err := r.checkQuery(sql, args...)
+	return &MockRows{results: rows, err: err}
+}
+
+func (r *SqlRecorder) CopyFrom(ctx context.Context, tableName pgx.Identifier, columnNames []string, rowSrc pgx.CopyFromSource) (int64, error) {
+	panic("should never be called")
+}
+
+func (r *SqlRecorder) CopyFromRows(rows [][]interface{}) pgx.CopyFromSource {
+	panic("should never be called")
+}
+
+func (r *SqlRecorder) NewBatch() pgxconn.PgxBatch {
+	return &MockBatch{}
+}
+
+func (r *SqlRecorder) SendBatch(ctx context.Context, b pgxconn.PgxBatch) (pgx.BatchResults, error) {
+	r.lock.Lock()
+	defer r.lock.Unlock()
+	batch := b.(*MockBatch)
+
+	start := r.nextQuery
+	for _, q := range batch.items {
+		_, _ = r.checkQuery(q.query, q.arguments...)
+	}
+	// TODO switch to q.query[] subslice
+	return &MockBatchResult{queries: r.queries[start:r.nextQuery]}, nil
+}
+
+func (r *SqlRecorder) checkQuery(sql string, args ...interface{}) (RowResults, error) {
+	idx := r.nextQuery
+	if idx >= len(r.queries) {
+		r.t.Errorf("@ %d extra query: %s", idx, sql)
+		return nil, fmt.Errorf("extra query")
+	}
+	row := r.queries[idx]
+	r.nextQuery += 1
+	if sql != row.sql {
+		r.t.Errorf("@ %d unexpected query:\ngot:\n\t%s\nexpected:\n\t%s", idx, sql, row.sql)
+	}
+	if !reflect.DeepEqual(args, row.args) {
+		r.t.Errorf("@ %d unexpected query args for\n\t%s\ngot:\n\t%#v\nexpected:\n\t%#v", idx, sql, args, row.args)
+	}
+	return row.results, row.err
+}
+
+type batchItem struct {
+	query     string
+	arguments []interface{}
+}
+
+// Batch queries are a way of bundling multiple queries together to avoid
+// unnecessary network round trips.
+type MockBatch struct {
+	items []batchItem
+}
+
+func (b *MockBatch) Queue(query string, arguments ...interface{}) {
+	b.items = append(b.items, batchItem{
+		query:     query,
+		arguments: arguments,
+	})
+}
+
+type MockBatchResult struct {
+	idx     int
+	queries []SqlQuery
+	t       *testing.T
+}
+
+// Exec reads the results from the next query in the batch as if the query has been sent with Conn.Exec.
+func (m *MockBatchResult) Exec() (pgconn.CommandTag, error) {
+	defer func() { m.idx++ }()
+
+	q := m.queries[m.idx]
+
+	if len(q.results) == 0 {
+		return nil, q.err
+	}
+	if len(q.results) != 1 {
+		m.t.Errorf("mock exec: too many return rows %v\n in batch Exec\n %+v", q.results, q)
+		return nil, q.err
+	}
+	if len(q.results[0]) != 1 {
+		m.t.Errorf("mock exec: too many return values %v\n in batch Exec\n %+v", q.results, q)
+		return nil, q.err
+	}
+
+	return q.results[0][0].(pgconn.CommandTag), q.err
+}
+
+// Query reads the results from the next query in the batch as if the query has been sent with Conn.Query.
+func (m *MockBatchResult) Query() (pgx.Rows, error) {
+	defer func() { m.idx++ }()
+
+	q := m.queries[m.idx]
+	return &MockRows{results: q.results, noNext: false}, q.err
+}
+
+// Close closes the batch operation. This must be called before the underlying connection can be used again. Any error
+// that occurred during a batch operation may have made it impossible to resyncronize the connection with the server.
+// In this case the underlying connection will have been closed.
+func (m *MockBatchResult) Close() error {
+	return nil
+}
+
+// QueryRow reads the results from the next query in the batch as if the query has been sent with Conn.QueryRow.
+func (m *MockBatchResult) QueryRow() pgx.Row {
+	defer func() { m.idx++ }()
+	q := m.queries[m.idx]
+	return &MockRows{results: q.results, err: q.err, noNext: false}
+}
+
+type MockRows struct {
+	idx     int
+	noNext  bool
+	results RowResults
+	err     error
+}
+
+// Close closes the rows, making the connection ready for use again. It is safe
+// to call Close after rows is already closed.
+func (m *MockRows) Close() {
+}
+
+// Err returns any error that occurred while reading.
+func (m *MockRows) Err() error {
+	return m.err
+}
+
+// CommandTag returns the command tag from this query. It is only available after Rows is closed.
+func (m *MockRows) CommandTag() pgconn.CommandTag {
+	panic("not implemented")
+}
+
+func (m *MockRows) FieldDescriptions() []pgproto3.FieldDescription {
+	panic("not implemented")
+}
+
+// Next prepares the next row for reading. It returns true if there is another
+// row and false if no more rows are available. It automatically closes rows
+// when all rows are read.
+func (m *MockRows) Next() bool {
+	return !m.noNext && m.idx < len(m.results)
+}
+
+// Scan reads the values from the current row into dest values positionally.
+// dest can include pointers to core types, values implementing the Scanner
+// interface, []byte, and nil. []byte will skip the decoding process and directly
+// copy the raw bytes received from PostgreSQL. nil will skip the value entirely.
+func (m *MockRows) Scan(dest ...interface{}) error {
+	defer func() { m.idx++ }()
+
+	if m.err != nil {
+		return m.err
+	}
+
+	if m.idx >= len(m.results) {
+		return fmt.Errorf("mock scanning error, no more results in batch: got %d wanted %d", m.idx, len(m.results))
+	}
+
+	if len(dest) > len(m.results[m.idx]) {
+		return fmt.Errorf("mock scanning error, missing results for scanning: got %d %#v\nwanted %d",
+			len(m.results[m.idx]),
+			m.results[m.idx],
+			len(dest),
+		)
+	}
+
+	for i := range dest {
+		switch s := m.results[m.idx][i].(type) {
+		case []time.Time:
+			if d, ok := dest[i].(*[]time.Time); ok {
+				*d = s
+			} else if d, ok := dest[i].(*pgtype.TimestamptzArray); ok {
+				*d = pgtype.TimestamptzArray{
+					Elements: make([]pgtype.Timestamptz, len(s)),
+				}
+				for i := range s {
+					d.Elements[i] = pgtype.Timestamptz{
+						Time: s[i],
+					}
+				}
+			}
+		case []float64:
+			if d, ok := dest[i].(*[]float64); ok {
+				*d = s
+			} else if d, ok := dest[i].(*pgtype.Float8Array); ok {
+				*d = pgtype.Float8Array{
+					Elements: make([]pgtype.Float8, len(s)),
+				}
+				for i := range s {
+					d.Elements[i] = pgtype.Float8{
+						Float: s[i],
+					}
+				}
+			}
+		case []int64:
+			if d, ok := dest[i].(*[]int64); ok {
+				*d = s
+				continue
+			}
+			if d, ok := dest[i].(*[]SeriesID); ok {
+				for _, id := range s {
+					*d = append(*d, SeriesID(id))
+				}
+				continue
+			}
+			return fmt.Errorf("wrong value type []int64")
+		case []string:
+			if d, ok := dest[i].(*[]string); ok {
+				*d = s
+			}
+		case time.Time:
+			if d, ok := dest[i].(*time.Time); ok {
+				*d = s
+			}
+		case float64:
+			if _, ok := dest[i].(float64); !ok {
+				return fmt.Errorf("wrong value type float64")
+			}
+			dv := reflect.ValueOf(dest[i])
+			dvp := reflect.Indirect(dv)
+			dvp.SetFloat(float64(m.results[m.idx][i].(float64)))
+		case int:
+			if _, ok := dest[i].(int); !ok {
+				return fmt.Errorf("wrong value type int")
+			}
+			dv := reflect.ValueOf(dest[i])
+			dvp := reflect.Indirect(dv)
+			dvp.SetInt(int64(m.results[m.idx][i].(int32)))
+		case int32:
+			if _, ok := dest[i].(int32); !ok {
+				return fmt.Errorf("wrong value type int32")
+			}
+			dv := reflect.ValueOf(dest[i])
+			dvp := reflect.Indirect(dv)
+			dvp.SetInt(int64(m.results[m.idx][i].(int32)))
+		case uint64:
+			if _, ok := dest[i].(uint64); !ok {
+				return fmt.Errorf("wrong value type uint64")
+			}
+			dv := reflect.ValueOf(dest[i])
+			dvp := reflect.Indirect(dv)
+			dvp.SetUint(m.results[m.idx][i].(uint64))
+		case int64:
+			_, ok1 := dest[i].(*int64)
+			_, ok2 := dest[i].(*SeriesID)
+			if !ok1 && !ok2 {
+				return fmt.Errorf("wrong value type int64")
+			}
+			dv := reflect.ValueOf(dest[i])
+			dvp := reflect.Indirect(dv)
+			dvp.SetInt(m.results[m.idx][i].(int64))
+		case string:
+			if _, ok := dest[i].(*string); !ok {
+				return fmt.Errorf("wrong value type string")
+			}
+			dv := reflect.ValueOf(dest[i])
+			dvp := reflect.Indirect(dv)
+			dvp.SetString(m.results[m.idx][i].(string))
+		}
+	}
+
+	return nil
+}
+
+// Values returns the decoded row values.
+func (m *MockRows) Values() ([]interface{}, error) {
+	panic("not implemented")
+}
+
+// RawValues returns the unparsed bytes of the row values. The returned [][]byte is only valid until the next Next
+// call or the Rows is closed. However, the underlying byte data is safe to retain a reference to and mutate.
+func (m *MockRows) RawValues() [][]byte {
+	panic("not implemented")
+}
+
+type MockMetricCache struct {
+	metricCache  map[string]string
+	getMetricErr error
+	setMetricErr error
+}
+
+func (m *MockMetricCache) Len() int {
+	return len(m.metricCache)
+}
+
+func (m *MockMetricCache) Cap() int {
+	return len(m.metricCache)
+}
+
+func (m *MockMetricCache) Get(metric string) (string, error) {
+	if m.getMetricErr != nil {
+		return "", m.getMetricErr
+	}
+
+	val, ok := m.metricCache[metric]
+	if !ok {
+		return "", ErrEntryNotFound
+	}
+
+	return val, nil
+}
+
+func (m *MockMetricCache) Set(metric string, tableName string) error {
+	m.metricCache[metric] = tableName
+	return m.setMetricErr
+}
+
+type MockCache struct {
+	seriesCache  map[string]SeriesID
+	getSeriesErr error
+	setSeriesErr error
+}
+
+var _ SeriesCache = (*MockCache)(nil)
+
+func (m *MockCache) GetSeries(lset Labels) (SeriesID, error) {
+	if m.getSeriesErr != nil {
+		return 0, m.getSeriesErr
+	}
+
+	val, ok := m.seriesCache[lset.String()]
+	if !ok {
+		return 0, cache.ErrEntryNotFound
+	}
+
+	return val, nil
+}
+
+func (m *MockCache) SetSeries(lset Labels, id SeriesID) error {
+	m.seriesCache[lset.String()] = id
+	return m.setSeriesErr
+}
+
+func (m *MockCache) NumElements() int {
+	return len(m.seriesCache)
+}
+
+func (m *MockCache) Capacity() int {
+	return len(m.seriesCache)
+}
+
+type MockInserter struct {
+	InsertedSeries  map[string]SeriesID
+	InsertedData    []map[string][]SamplesInfo
+	InsertSeriesErr error
+	InsertDataErr   error
+}
+
+func (m *MockInserter) Close() {
+
+}
+
+func (m *MockInserter) InsertNewData(rows map[string][]SamplesInfo) (uint64, error) {
+	return m.InsertData(rows)
+}
+
+func (m *MockInserter) CompleteMetricCreation() error {
+	return nil
+}
+
+func (m *MockInserter) InsertData(rows map[string][]SamplesInfo) (uint64, error) {
+	for _, v := range rows {
+		for i, si := range v {
+			id, ok := m.InsertedSeries[si.Labels.String()]
+			if !ok {
+				id = SeriesID(len(m.InsertedSeries))
+				m.InsertedSeries[si.Labels.String()] = id
+			}
+			v[i].SeriesID = id
+		}
+	}
+	if m.InsertSeriesErr != nil {
+		return 0, m.InsertSeriesErr
+	}
+	m.InsertedData = append(m.InsertedData, rows)
+	ret := 0
+	for _, data := range rows {
+		for _, si := range data {
+			ret += len(si.Samples)
+		}
+	}
+	if m.InsertDataErr != nil {
+		ret = 0
+	}
+	return uint64(ret), m.InsertDataErr
+}
